modifications suggested by iHireTechnology system
not quite done yet:

You included an objective statement in your resume. This is an outdated
practice and should be replaced with a compelling summary paragraph.
Your resume appears to be lacking an Executive Summary. We recommend
adding an Executive Summary for a strong, interview winning resume. If
your resume has a summary section but our scorer is unable to recognize
it, consider adding “Summary” as a section title or lead in.
Your resume appears to be lacking Achievements. We recommend adding
Achievements for a strong, interview winning resume.